Move random number functions to file and make docs
I am beginning to refactor code. To make it so we can reuse the random
number functions (e.g. the ones that generate arrays of random ints and
bytes), I am separating all of those functions out to a file.

I also make documentation for all of them in a header file and setup
doxygen to document the code. Yay for documentation!