Add Bayesian model for caching
I created a new test and Rscript which tests multiple sizes and bucket
sizes simultaneously. I built a model in STAN, a probabilistic modeling
language, with which I could estimate simultaneously the size of the
L1 and L2 cache as well as access times for those caches. They seem to
be overestimates right now, but I think a model that includes more
caches and memory will be more effective when I create that. But this
is a good start on creating an informative model to solve this problem.